Pie chart
- Analysis chart re-organised
- Pie chart with redundancy distribution added
