+++ JAG3D/resources/i18n/jag3d.properties
@@ -1902,24 +1902,46 @@ TableRowHighlightDialog.message.error.save.exception.header  = Error, could not 
 TableRowHighlightDialog.message.error.save.exception.message = An exception has occurred during database transaction.
 
 
-ResidualAnalysisDialog.title                        = Analysis charts
-ResidualAnalysisDialog.chart.histogram.header       = Histogram of normalized residuals
-ResidualAnalysisDialog.chart.histogram.axis.x.label = Normalized residuals
-ResidualAnalysisDialog.chart.histogram.axis.y.label = Probability density
+AnalysisChartsDialog.title                        = Analysis charts
+AnalysisChartsDialog.chart.header                 = Analysis charts
+AnalysisChartsDialog.chart.type.residuals.header  = Histogram of normalized residuals
+AnalysisChartsDialog.chart.type.redundancy.header = Pie chart of redundancy
+AnalysisChartsDialog.chart.type.tooltip           = Select analysis chart type
 
-ResidualAnalysisDialog.observationtype.terrestrial.tooltip                   = Set observational residual type
-ResidualAnalysisDialog.observationtype.terrestrial.all.label                 = All terrestrial observations
-ResidualAnalysisDialog.observationtype.terrestrial.leveling.label            = Leveling
-ResidualAnalysisDialog.observationtype.terrestrial.direction.label           = Directions
-ResidualAnalysisDialog.observationtype.terrestrial.horizontal_distance.label = Horizontal distances
-ResidualAnalysisDialog.observationtype.terrestrial.slope_distance.label      = Slope distances
-ResidualAnalysisDialog.observationtype.terrestrial.zenith_angle.label        = Zenith angles
 
-ResidualAnalysisDialog.histogram.probability_density_function.label   = Gaussian probability density function
-ResidualAnalysisDialog.histogram.probability_density_function.tooltip = If selected, the probability density function of the standard normal distribution is estimated
-ResidualAnalysisDialog.histogram.kernel_density_estimation.label      = Kernel density estimation
-ResidualAnalysisDialog.histogram.kernel_density_estimation.tooltip    = If selected, the probability density function is estimated by a kernel density estimation
+AnalysisChartTypeListViewCell.residuals.label  = Residual analysis
+AnalysisChartTypeListViewCell.redundancy.label = Redundancy
 
-ResidualAnalysisDialog.message.error.load.exception.title   = Unexpected SQL-Error
-ResidualAnalysisDialog.message.error.load.exception.header  = Error, could not load data from database.
-ResidualAnalysisDialog.message.error.load.exception.message = An exception has occurred during database transaction.
+
+UIResidualAnalysisChart.chart.histogram.axis.x.label = Normalized residuals
+UIResidualAnalysisChart.chart.histogram.axis.y.label = Probability density
+
+UIResidualAnalysisChart.observationtype.terrestrial.tooltip                   = Set observational residual type
+UIResidualAnalysisChart.observationtype.terrestrial.all.label                 = All terrestrial observations
+UIResidualAnalysisChart.observationtype.terrestrial.leveling.label            = Leveling
+UIResidualAnalysisChart.observationtype.terrestrial.direction.label           = Directions
+UIResidualAnalysisChart.observationtype.terrestrial.horizontal_distance.label = Horizontal distances
+UIResidualAnalysisChart.observationtype.terrestrial.slope_distance.label      = Slope distances
+UIResidualAnalysisChart.observationtype.terrestrial.zenith_angle.label        = Zenith angles
+
+UIResidualAnalysisChart.histogram.probability_density_function.label   = Gaussian probability density function
+UIResidualAnalysisChart.histogram.probability_density_function.tooltip = If selected, the probability density function of the standard normal distribution is estimated
+UIResidualAnalysisChart.histogram.kernel_density_estimation.label      = Kernel density estimation
+UIResidualAnalysisChart.histogram.kernel_density_estimation.tooltip    = If selected, the probability density function is estimated by a kernel density estimation
+
+UIResidualAnalysisChart.message.error.load.exception.title   = Unexpected SQL-Error
+UIResidualAnalysisChart.message.error.load.exception.header  = Error, could not load data from database.
+UIResidualAnalysisChart.message.error.load.exception.message = An exception has occurred during database transaction.
+
+
+UIRedundancyAnalysisChart.observationtype.terrestrial.tooltip                   = Set observational redundancy type
+UIRedundancyAnalysisChart.observationtype.terrestrial.all.label                 = All terrestrial observations
+UIRedundancyAnalysisChart.observationtype.terrestrial.leveling.label            = Leveling
+UIRedundancyAnalysisChart.observationtype.terrestrial.direction.label           = Directions
+UIRedundancyAnalysisChart.observationtype.terrestrial.horizontal_distance.label = Horizontal distances
+UIRedundancyAnalysisChart.observationtype.terrestrial.slope_distance.label      = Slope distances
+UIRedundancyAnalysisChart.observationtype.terrestrial.zenith_angle.label        = Zenith angles
+
+UIRedundancyAnalysisChart.message.error.load.exception.title   = Unexpected SQL-Error
+UIRedundancyAnalysisChart.message.error.load.exception.header  = Error, could not load data from database.
+UIRedundancyAnalysisChart.message.error.load.exception.message = An exception has occurred during database transaction.
