+++ JAG3D/resources/i18n/jag3d_sr_BA.properties
@@ -1591,9 +1591,17 @@ TableRowHighlightDialog.message.error.save.exception.header  = Gre\u0161ka, nije
 TableRowHighlightDialog.message.error.save.exception.message = Do\u0161lo je do neo\u010Dekivane gre\u0161ke prilikom spremanja u bazu.
 
 
-ResidualAnalysisDialog.observationtype.terrestrial.all.label                 = Terestri\u010Dka mjerenja
-ResidualAnalysisDialog.observationtype.terrestrial.leveling.label            = Nivelmanska mjerenja
-ResidualAnalysisDialog.observationtype.terrestrial.direction.label           = Pravci
-ResidualAnalysisDialog.observationtype.terrestrial.horizontal_distance.label = Horizontalne du\u017Eine
-ResidualAnalysisDialog.observationtype.terrestrial.slope_distance.label      = Kose du\u017Eine
-ResidualAnalysisDialog.observationtype.terrestrial.zenith_angle.label        = Zenitni uglovi
+UIResidualAnalysisChart.observationtype.terrestrial.all.label                 = Terestri\u010Dka mjerenja
+UIResidualAnalysisChart.observationtype.terrestrial.leveling.label            = Nivelmanska mjerenja
+UIResidualAnalysisChart.observationtype.terrestrial.direction.label           = Pravci
+UIResidualAnalysisChart.observationtype.terrestrial.horizontal_distance.label = Horizontalne du\u017Eine
+UIResidualAnalysisChart.observationtype.terrestrial.slope_distance.label      = Kose du\u017Eine
+UIResidualAnalysisChart.observationtype.terrestrial.zenith_angle.label        = Zenitni uglovi
+
+
+UIRedundancyAnalysisChart.observationtype.terrestrial.all.label                 = Terestri\u010Dka mjerenja
+UIRedundancyAnalysisChart.observationtype.terrestrial.leveling.label            = Nivelmanska mjerenja
+UIRedundancyAnalysisChart.observationtype.terrestrial.direction.label           = Pravci
+UIRedundancyAnalysisChart.observationtype.terrestrial.horizontal_distance.label = Horizontalne du\u017Eine
+UIRedundancyAnalysisChart.observationtype.terrestrial.slope_distance.label      = Kose du\u017Eine
+UIRedundancyAnalysisChart.observationtype.terrestrial.zenith_angle.label        = Zenitni uglovi
